requirements.txt
Flask==3.0.3
gunicorn==23.0.0
joblib==1.4.2
Pillow==11.0.0
scikit-learn==1.5.2
numpy==2.1.3
certifi==2024.8.30

